MERGE: Merged trunk into ptx 2.1 branch up to r597.  ENH: Added new .pragma statement to PTX IR.
Former-commit-id: 6ef9313b8ae7301477995514facb49feccb6c754

+++ branches/ocelot-ptx-2.1/ocelot/executive/interface/LLVMContext.h
@@ -20,9 +20,9 @@ namespace executive
 			class Dimension
 			{
 				public:
-					unsigned short x;
-					unsigned short y;
-					unsigned short z;
+					unsigned int x;
+					unsigned int y;
+					unsigned int z;
 			};
 	
 		public:
